feat(cmake): enable serialization tests in CI_LITE (#590)
* feat(cmake): enable serialization tests in CI_LITE

Add UNIVERSAL_BUILD_LINEAR_ALGEBRA_DATA to the CI_LITE cascade so
serialization tests (data_serialization, data_typed_serialization,
data_translation) are built and run in CI on every PR.

Previously these tests only ran with UNIVERSAL_BUILD_ALL, which is
not used in CI. This meant serialization regressions (like the
lns::fraction() stub in #509) were invisible to CI.

Resolves #586


* fix(blas): add value verification to serialization round-trip tests

Tests 6-8 (lns, dbns, integer) now verify bit-exact fidelity by
comparing the serialized text after a save->restore->re-save cycle.
If the re-serialized text differs from the original, the round-trip
lost data.


* fix(blas): guard value comparison on successful restore

Make the re-save/compare check conditional on restore succeeding
so a failed restore doesn't double-count as a value mismatch.
Applied to lns, dbns, and integer round-trip tests.


* fix(fixpnt): silence uninitialized warning in numeric_limits

fixpnt's epsilon(), round_error(), and denorm_min() declared a local
FixedPoint without initialization, then called setbit() on it. GCC
warned that the underlying blockbinary storage was uninitialized.

Fix: value-initialize with {} and call clear() before setbit() to
ensure all bits are zero before setting the ULP bit.


* fix(blockbinary): silence GCC -Warray-bounds false positive in setbit

GCC at -O3 produces false-positive -Warray-bounds warnings when
template inlining causes it to confuse blockbinary instantiations
across fixpnt's convert() -> setbit() chain. The existing guard
(blockIndex < nrBlocks) is correct but GCC's analyzer loses track
of which instantiation's array is being accessed.

Fix: add early return (i >= nbits) before the blockIndex calculation.
This gives GCC a stronger hint that out-of-range indices are
impossible, silencing the warning without changing behavior.


* fix(blockbinary): suppress GCC false-positive -Warray-bounds in setbit

GCC -O3 produces false-positive -Warray-bounds warnings when inlining
setbit() across blockbinary instantiations with different nbits (e.g.,
blockbinary<129> inlined into blockbinary<128> context via assign(),
maxneg(), or operator>>=). The i>=nbits and blockIndex<nrBlocks guards
are correct but GCC's interprocedural analysis loses track of which
instantiation's _block array is being referenced.

Add targeted #pragma GCC diagnostic push/pop around the array access
to suppress the false positive. The pragma is GCC-only (guarded by
__GNUC__ && !__clang__) since clang does not have this issue.


* fix(blockbinary): also suppress GCC -Wstringop-overflow false positive

GCC emits both -Warray-bounds and -Wstringop-overflow for the same
cross-instantiation inlining issue in setbit(). Add suppression for
-Wstringop-overflow (available since GCC 12) alongside -Warray-bounds.


* fix(rational): value-initialize locals to silence uninitialized warning

GCC warns that rational<8> locals may be used uninitialized when
the default constructor (= default) leaves blockbinary members
uninitialized. Value-initialize with {} to zero the storage.


* fix(rational): value-initialize remaining assignment test locals

Same uninitialized warning fix for hexadecimal and decimal rational
assignment tests.


* fix(unum2): resolve sign-compare warnings in lattice and SORN code

Change lattice members _N, _N_half, _N_quarter, _MASK from int64_t
to uint64_t (they are always non-negative, computed from sizeof...).

Fix loop variables and sentinels in unum2_impl.hpp and op_matrix.hpp
to use consistent unsigned types, eliminating all -Wsign-compare
warnings.

Replace int64_t sentinel left_bound=-1 with bool has_left_bound flag.


* fix(areal): suppress GCC false-positive array-bounds in set()

Same GCC -O3 cross-instantiation inlining false positive as
blockbinary::setbit(). The areal::set() has its own _block array
and the i<nbits guard is correct, but GCC confuses areal<16> and
areal<32> instantiations.

Apply the same pragma pattern: early return for i>=nbits, explicit
blockIndex<nrBlocks check, and GCC-only diagnostic suppression.


---------

+++ include/sw/universal/number/unum2/op_matrix.hpp
@@ -38,8 +38,8 @@ public:
             throw;  // rethrow
         }
         
-        for(int i = 0; i < OP_MATRIX_TOTAL_SUPPORTED_OPS; i++) {
-            for(int j = 0; j < N * N; j++) 
+        for(size_t i = 0; i < OP_MATRIX_TOTAL_SUPPORTED_OPS; i++) {
+            for(size_t j = 0; j < N * N; j++)
                 _present_flags[i][j] = false;
         }
     }
